методы по сохранению пользовательских данных вынесены в новый Smart class, навешаны обработчики на выход без сохранения формы
